Adds the "choose a phone number" flow to the UI on first load
When the UI is first loaded, the user will have to choose a phone number
for the dev-phone.

They are warned if the number they are choosing already has voice or SMS
webhooks configured, but are free to ignore those warnings.

Once a phone number is chosen, it's set globally for the app. The only
way to change this choice is to refresh the page.