Fixes shutdown behaviour
Because the function passed to `process.on('SIGINT'...)` is
in its own scope we couldn't call this.someFunction directly.

This commit creates a function which closes over `this` which
can then be called inside the `process.on` function and will
take `this` with it so we can tidy up the user's account when
the plugin stops.