Use longer-lived AccessTokens by default
Bump up the TTL (time-to-live) value for Access Tokens we create. The
default is 3600 seconds (one hour), this changes it to 24*60*60 seconds
(24 hours).

This is a band-aid style fix because the correct solution is to be
hooking into one of the client callbacks for `tokenAboutToExpire` and
generating a new one transparently to the user. But that might be quite
a big fix, and perhaps we can live with this as a known limitation for
Tweek.

See Issue #51